+++ lib/query_owl/engine.rb
@@ -13,22 +13,7 @@ module QueryOwl
     end
 
     initializer "query_owl.request_tracking" do |app|
-      app.middleware.use(Class.new do
-        def initialize(rack_app)
-          @app = rack_app
-        end
-
-        def call(env)
-          return @app.call(env) unless QueryOwl.config.enabled
-
-          QueryTracker.start!
-          @app.call(env)
-        ensure
-          queries = QueryTracker.stop!
-          events  = Detector.detect_n_plus_one(queries) + Detector.detect_slow_queries(queries)
-          Logger.log_events(events)
-        end
-      end)
+      app.middleware.use(Middleware)
     end
   end
 end